App component split up between 5 new components but functionality is not working
