+++ react/src/components/MemoryType.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const MemoryType = props => {
+
+  return(
+    <div className="container monologue-text-parts">
+      <button className="back-monologue btn waves-effect waves-teal orange darken-3" onClick={props.memoryTesterClick}>
+        {`Back To Your Monologue`}
+      </button><br/>
+        <input className="text-box" type="text"
+        onChange={props.update} onKeyDown={props.handleSpacePress} />
+      <p className="white-text text-state">{props.memory}</p>
+    </div>
+  )
+}
+
+export default MemoryType;
